Edited GET & POST tasks route to use Family model
Edited the GET tasks route to use the Family model, instead
of the Task model, to find tasks associated with a familyCode.

Edited the POST tasks route. After a task is created using
the Task model, the task is added to the `families` collection
using the Family model before sending the task back as JSON.