Check if `tasksFinalized` is true before finalizing tasksFinalized
Before finalizing tasks, check if the `tasksFinalized` key
in the Family model is already True. This is to prevent the backend
data from unintentionally changing. If `tasksFinalized` is True,
send an error that `tasksFinalized` is already True. If False,
set `tasksFinalized` to True and save the document.

Family.findOneAndUpdate() was passed a third argument, an options
object { new: true, upsert: true, }.
`new` if true, returns the modified Family document rather
than the original. `upsert` if true, creates the Family object
if it doesn’t exist.