Check if `tasksFinalized` is true before Creating Task
Before creating a new Task, make sure the `tasksFinalized` key
in the Family model is False (not True). This is to prevent adding
new Tasks to the tasks list, once the Parent has finalized
the tasks list. If `tasksFinalized` is True, send an error message
`tasksFinalized` is already True. If not True, proceed
to create a new Task.

Family.findOneAndUpdate() was passed a third argument, an options
object { new: true, upsert: true, }.
`new` if true, returns the modified Family document rather
than the original. `upsert` if true, creates the Family object
if it doesn’t exist.