Check if `isTasksFinalized` is true before Updating Task
Before Updating an existing Task, make sure the `tasksFinalized` key
in the Family model is False (not True). This is to prevent updating
existing Tasks in the tasks list, once the Parent has finalized
the tasks list. If `tasksFinalized` is True, send an error message
`tasksFinalized` is already True. If not True, proceed
to update an existing Task.