Check if `tasksFinalized` is true before Deleting Task
Before Deleting a Task, make sure the `tasksFinalized` key
in the Family model is False (not True). This is to prevent deleting
Tasks in the tasks list, once the Parent has finalized
the tasks list. If `tasksFinalized` is True, send an error message
`tasksFinalized` is already True. If not True, proceed
to delete a Task.