Check if `tasksFinalized` is true before Updating Prize
Before Updating a Prize, make sure the `tasksFinalized` key
in the Family model is False (not True). This is to prevent updating
a Prize, once the Parent has finalized the tasks list.
If `tasksFinalized` is True, send an error message `tasksFinalized`
is already True. If not True, proceed to update a Prize.

Family.findOneAndUpdate() was passed a third argument, an options
object { new: true, upsert: true, }.
`new` if true, returns the modified Family document rather
than the original. `upsert` if true, creates the Family object
if it doesn’t exist.

NOTE: The PUT req to the `/api/prizes` URL might not be made or needed
based on the `createOrUpdatePrize` thunk on the Client side.
The thunk makes a POST fetch request to `/api/prizes`.