Reset the Child's `tasksReadyForReview` field when the tasks list is reset
In the Family document, the following three field values are reset
when the Parent clicks the `Reset Tasks List` button:
* tasksFinalized: false,
* currentTasks: [],
* currentPrize: null

This reset affects all Child users for a given Family.

In addition to resetting the three field values in the Family
document, the individual Child user's `tasksReadyForReview`
field is reset to False.