Edited tasks-router
* Required Family model.
* In the beforeEach hook, after first creating a new user,
call seedTasks to generate an array of mock tasks. Next take
the tasks array and map to save their Id's. Next using the Family model,
update the family document by inserting the currentTasks field
with the array of task Id's, as well as the familyCode name.
The `upsert` command will create the document with the updated info
if it doesn't exist.