Upgrade mongoose from v5 to v6
The `task-treasure-db` database on the M0 Atlas cluster will
automatically be upgraded to v5.0. In order to use MongoDB v5.0,
the Node.js driver needs to be upgraded to v4.x.

Mongoose relies on the MongoDB Node.js driver to talk to MongoDB.
Mongoose needs to be upgraded to v6.0 so it can use the Node.js v4.x
driver, in order to work with MongoDB 5.0.

Note: the `task-treasure-db` database is part of the
`development-task-treasure-db` project on MongoDB Atlas.